Added frmBakedGoods and frmOrders
What I've done

1) Created frmOrders and frmBakedGood screens.

2) Created the inheritance relationship with webform1.

3) Committed changes to GitHub repository.

What I'm doing Next.

There needs to be an orderitem user input screen that can be

launched from the frmOrders screen that allows the customer to add a new
lineitem to an existing order.

Problems:

The creation/inheritance of the aforementioned forms has created a null
exception pointer error in the Customers class.

It seems to be related to the dynamic creation of the controls in that
class (see screeshot in PCC discussion group standup #4 attached file).

Take a look at the problem and let me know what you think may be the
cause.