Commented out some code to get it to run
By commenting out the code inside the Page_Load, FillBakedItemList, and
FillOrders() of Customers.aspx.vb, ; as well as commenting out the  the
code relating to the lstCustomerOrders control in the
customers.aspx.designer.vb,
I was able to run the program.
Take a look at the screens and let me know if you think this matches the
feel of what you designed.